Updates Default Netplay Buffer to 16 (server)
Goes along with this commit here: https://github.com/Tinob/Ishiiruka/pull/54/commits/cdf9e75a9ea3f50975a28a56e38f07a0c22967e2

This is to make the default pad buffer for netplay "16" instead of "5".

The reason for this is with the new netplay buffer system, each buffer is 25% of what it was before - it takes 4x the buffer now for it to be equal. 16 seems to be the optimal value based on avg ping.